Split main.js into focused modules (#109)
Extract logical groups from the 3627-line monolithic main.js into
5 focused modules:

- paths.js: shared path constants and config flags
- daemon-client.js: PTY daemon socket communication
- session-discovery.js: session state detection, caching, origin tagging
- pool-manager.js: pool lifecycle, offload/archive, terminal helpers
- api-handlers.js: shared IPC/API handler registry, API-only handlers

main.js remains as the orchestrator (~350 lines): Electron lifecycle,
window creation, IPC wiring, menu building, file watching, intervals.

Circular dependency between session-discovery and pool-manager solved
with lazy require pattern (pool-manager lazily requires session-discovery).
